requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
chardet==5.2.0
crispy-bootstrap5==2024.10
distro==1.9.0
Django==5.1.6
django-crispy-forms==2.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.1
openai==1.63.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-docx==1.1.2
python-dotenv==1.0.1
reportlab==4.3.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
gunicorn==21.2.0
